Move quit test to header
