+++ frontend/src/components/commons/TestFooter.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import LOCALIZE from "../../text_resources";
-import { Navbar, Nav } from "react-bootstrap";
+import { Navbar } from "react-bootstrap";
 
 const styles = {
   footer: {
@@ -15,7 +15,6 @@ class TestFooter extends Component {
   static propTypes = {
     startTest: PropTypes.func,
     submitTest: PropTypes.func,
-    quitTest: PropTypes.func,
     testIsStarted: PropTypes.bool.isRequired
   };
 
@@ -23,20 +22,6 @@ class TestFooter extends Component {
     return (
       <div>
         <Navbar fixed="bottom" style={styles.footer}>
-          <Nav className="mr-auto">
-            {this.props.testIsStarted && (
-              <Navbar.Text>
-                <button
-                  id="unit-test-quit-btn"
-                  type="button"
-                  className="btn btn-danger"
-                  onClick={this.props.quitTest}
-                >
-                  {LOCALIZE.commons.quitTest}
-                </button>
-              </Navbar.Text>
-            )}
-          </Nav>
           {!this.props.testIsStarted && (
             <Navbar.Text className="justify-content-end">
               <button
