+++ frontend/src/components/eMIB/Emib.jsx
@@ -7,8 +7,7 @@ import EmibTabs from "./EmibTabs";
 import TestFooter from "../commons/TestFooter";
 import LOCALIZE from "../../text_resources";
 import ContentContainer from "../commons/ContentContainer";
-import PopupBox, { BUTTON_TYPE, BUTTON_STATE } from "../commons/PopupBox";
-import SystemMessage, { MESSAGE_TYPE } from "../commons/SystemMessage";
+import PopupBox, { BUTTON_TYPE } from "../commons/PopupBox";
 import { activateTest, deactivateTest } from "../../modules/TestStatusRedux";
 import ConfirmStartTest from "../commons/ConfirmStartTest";
 import EmibIntroductionPage from "./EmibIntroductionPage";
@@ -20,69 +19,29 @@ const PAGES = {
   confirm: "confirm"
 };
 
-const styles = {
-  hr: {
-    width: "100%",
-    borderTop: "2px solid #96a8b2",
-    margin: "16px 0 16px 0"
-  },
-  checkboxZone: {
-    paddingTop: 8
-  },
-  startTestBtn: {
-    textAlign: "center",
-    marginTop: 32
-  }
-};
-
-const quitConditions = () => {
-  return [
-    { text: LOCALIZE.emibTest.testFooter.quitTestPopupBox.checkboxOne, checked: false },
-    { text: LOCALIZE.emibTest.testFooter.quitTestPopupBox.checkboxTwo, checked: false },
-    { text: LOCALIZE.emibTest.testFooter.quitTestPopupBox.checkboxThree, checked: false }
-  ];
-};
-
 class Emib extends Component {
   static propTypes = {
     // Provided by Redux
     activateTest: PropTypes.func.isRequired,
-    deactivateTest: PropTypes.func.isRequired
+    deactivateTest: PropTypes.func.isRequired,
+    curPage: PropTypes.string.isRequired
   };
 
   state = {
-    curPage: PAGES.preTest,
     currentTab: "instructions",
     disabledTabs: [1, 2],
     testIsStarted: false,
     showStartTestPopup: false,
-    showSubmitPopup: false,
-    showQuitPopup: false,
-    quitConditions: quitConditions()
-  };
-
-  changePage = () => {
-    switch (this.state.curPage) {
-      case PAGES.preTest:
-        // Move from instructions to starting the test.
-        this.setState({ curPage: PAGES.emibTabs });
-        // update redux to activate test
-        this.props.activateTest();
-        break;
-      case PAGES.emibTabs:
-        this.setState({ curPage: PAGES.confirm });
-        // update redux to de-activate test
-        this.props.deactivateTest();
-        break;
-      default:
-        this.setState({ curPage: PAGES.preTest });
-        break;
-    }
+    showSubmitPopup: false
   };
 
   // Within eMIB Tabs functions
   handleStartTest = () => {
     this.setState({ testIsStarted: true, disabledTabs: [], currentTab: "background" });
+  };
+
+  closePopup = () => {
+    this.setState({ showSubmitPopup: false, showQuitPopup: false });
   };
 
   switchTab = tabId => {
@@ -102,61 +61,32 @@ class Emib extends Component {
     this.setState({ showSubmitPopup: true });
   };
 
-  closePopup = () => {
-    this.setState({ showSubmitPopup: false, showQuitPopup: false });
-    //reset all checkbox states on close
-    this.resetCheckboxStates();
-  };
-
-  openQuitPopup = () => {
-    this.setState({ showQuitPopup: true });
-  };
-
-  resetCheckboxStates = () => {
-    this.setState({ quitConditions: quitConditions() });
-  };
-
-  toggleCheckbox = id => {
-    let updatedQuitConditions = Array.from(this.state.quitConditions);
-    updatedQuitConditions[id].checked = !updatedQuitConditions[id].checked;
-    this.setState({ quitConditions: updatedQuitConditions });
-  };
-
-  isChecked = currentCheckbox => {
-    return currentCheckbox.checked;
-  };
-
   render() {
-    const { quitConditions } = this.state;
-    const allChecked = quitConditions.every(this.isChecked);
-
-    const submitButtonState = allChecked ? BUTTON_STATE.enabled : BUTTON_STATE.disabled;
     return (
       <div className="app">
         <Helmet>
           <title>{LOCALIZE.titles.eMIB}</title>
         </Helmet>
-        {this.state.curPage === PAGES.emibTabs && (
+        {this.props.curPage === PAGES.emibTabs && (
           <EmibTabs
             currentTab={this.state.currentTab}
             switchTab={this.switchTab}
             disabledTabsArray={this.state.disabledTabs}
           />
         )}
-        {this.state.curPage !== PAGES.emibTabs && (
+        {this.props.curPage !== PAGES.emibTabs && (
           <ContentContainer hideBanner={false}>
-            {this.state.curPage === PAGES.preTest && (
-              <EmibIntroductionPage nextPage={this.changePage} />
+            {this.props.curPage === PAGES.preTest && (
+              <EmibIntroductionPage nextPage={this.props.activateTest} />
             )}
 
-            {this.state.curPage === PAGES.confirm && <Confirmation />}
+            {this.props.curPage === PAGES.confirm && <Confirmation />}
           </ContentContainer>
         )}
-        {this.state.curPage === PAGES.emibTabs && (
+        {this.props.curPage === PAGES.emibTabs && (
           <TestFooter
             startTest={this.openStartTestPopup}
             submitTest={this.openSubmitPopup}
-            quitTest={this.openQuitPopup}
             testIsStarted={this.state.testIsStarted}
           />
         )}
@@ -181,62 +111,7 @@ class Emib extends Component {
           leftButtonTitle={LOCALIZE.commons.cancel}
           rightButtonType={BUTTON_TYPE.primary}
           rightButtonTitle={LOCALIZE.commons.submitTestButton}
-          rightButtonAction={this.changePage}
-        />
-
-        <PopupBox
-          show={this.state.showQuitPopup}
-          handleClose={this.closePopup}
-          title={LOCALIZE.emibTest.testFooter.quitTestPopupBox.title}
-          description={
-            <div>
-              <div>
-                <SystemMessage
-                  messageType={MESSAGE_TYPE.error}
-                  title={LOCALIZE.emibTest.testFooter.quitTestPopupBox.warning.title}
-                  message={LOCALIZE.emibTest.testFooter.quitTestPopupBox.warning.message}
-                />
-              </div>
-              <p className="font-weight-bold">
-                {LOCALIZE.emibTest.testFooter.quitTestPopupBox.descriptionPart1}
-              </p>
-              <div>
-                {this.state.quitConditions.map((condition, id) => {
-                  return (
-                    <div
-                      key={id}
-                      className="custom-control custom-checkbox"
-                      style={styles.checkboxZone}
-                    >
-                      <input
-                        type="checkbox"
-                        className="custom-control-input"
-                        id={id}
-                        checked={condition.checked}
-                        onChange={() => this.toggleCheckbox(id)}
-                      />
-                      <label className="custom-control-label" htmlFor={id}>
-                        {condition.text}
-                      </label>
-                    </div>
-                  );
-                })}
-              </div>
-              <hr style={styles.hr} />
-              <p className="font-weight-bold">
-                {LOCALIZE.emibTest.testFooter.quitTestPopupBox.descriptionPart2}
-              </p>
-              <p className="font-weight-bold">
-                {LOCALIZE.emibTest.testFooter.quitTestPopupBox.descriptionPart3}
-              </p>
-            </div>
-          }
-          leftButtonType={BUTTON_TYPE.danger}
-          leftButtonTitle={LOCALIZE.commons.quitTest}
-          leftButtonAction={this.changePage}
-          leftButtonState={submitButtonState}
-          rightButtonType={BUTTON_TYPE.primary}
-          rightButtonTitle={LOCALIZE.commons.returnToTest}
+          rightButtonAction={this.props.deactivateTest}
         />
       </div>
     );
@@ -244,6 +119,12 @@ class Emib extends Component {
 }
 export { PAGES };
 export { Emib as UnconnectedEmib };
+
+const mapStateToProps = (state, ownProps) => {
+  return {
+    curPage: state.testStatus.currentPage
+  };
+};
 
 const mapDispatchToProps = dispatch =>
   bindActionCreators(
@@ -255,6 +136,6 @@ const mapDispatchToProps = dispatch =>
   );
 
 export default connect(
-  null,
+  mapStateToProps,
   mapDispatchToProps
 )(Emib);
